requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.35.0
jinja2>=3.1.6
python-multipart>=0.0.20
openai>=1.107.0
playwright>=1.58.0
httpx>=0.28.1
feedparser>=6.0.11
beautifulsoup4>=4.13.4
apscheduler>=3.11.0
python-dotenv>=1.1.1
pydantic>=2.11.7
python-docx>=0.8.12
python-docx>=1.1.2
